refactored respective serialisation methods
